# nrf52: nvmc: Erase page only if it is not empty